# swagger/draft.yaml
openapi: 3.0.3
info:
  title: rfiles
  description: rfiles api draft
  contact:
    email: ksi123456ab@gmail.com
  version: 0.0.1
tags:
  - name: objects
    description: files
paths:
  /objects:
    get:
      tags:
        - objects
      summary: List all objects
      description: List all objects the server stores
      parameters:
        - name: return
          in: query
          description: return mode
          schema:
            type: string
            default: md5
            enum:
              - md5
              - object
      responses:
        '200':
          description: object list
  /objects/{hash}:
    head:
      tags:
        - objects
      summary: Get metadata
      description: Get metadata of the file by its hash
      parameters:
        - name: hash
          in: path
          description: checksum of file
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File found
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: string
    get:
      tags:
        - objects
      summary: Download a file
      description: Download a file by its hash
      parameters:
        - name: hash
          in: path
          description: checksum of file
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File Found
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Not found
    put:
      tags:
        - objects
      summary: Upload a file
      description: Update a file by its hash
      parameters:
        - name: hash
          in: path
          description: checksum of file
          required: true
          schema:
            type: string
        - name: overwrite
          in: query
          description: Behavior if file already exists
          required: false
          schema:
            type: string
            default: error
            enum:
              - error
              - overwrite
              - skip
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
        required: true
      responses:
        '204':
          description: Successful operation (no content)
        '400':
          description: 'bad request or "mismatch-hash"'
      security:
        - petstore_auth:
            - write:pets
            - read:pets
    delete:
      tags:
        - objects
      summary: Delete a file
      description: Delete a file by its hash
      parameters:
        - name: hash
          in: path
          description: checksum of file
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successful operation
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        petId:
          type: integer
          format: int64
          example: 198772
        quantity:
          type: integer
          format: int32
          example: 7
        shipDate:
          type: string
          format: date-time
        status:
          type: string
          description: Order Status
          example: approved
          enum:
            - placed
            - approved
            - delivered
        complete:
          type: boolean
      xml:
        name: order
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 100000
        username:
          type: string
          example: fehguy
        address:
          type: array
          xml:
            name: addresses
            wrapped: true
          items:
            $ref: '#/components/schemas/Address'
      xml:
        name: customer
    Address:
      type: object
      properties:
        street:
          type: string
          example: 437 Lytton
        city:
          type: string
          example: Palo Alto
        state:
          type: string
          example: CA
        zip:
          type: string
          example: '94301'
      xml:
        name: address
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Dogs
      xml:
        name: category
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: '12345'
        phone:
          type: string
          example: '12345'
        userStatus:
          type: integer
          description: User Status
          format: int32
          example: 1
      xml:
        name: user
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: tag
    Pet:
      required:
        - name
        - photoUrls
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: doggie
        category:
          $ref: '#/components/schemas/Category'
        photoUrls:
          type: array
          xml:
            wrapped: true
          items:
            type: string
            xml:
              name: photoUrl
        tags:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Tag'
        status:
          type: string
          description: pet status in the store
          enum:
            - available
            - pending
            - sold
      xml:
        name: pet
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'
  requestBodies:
    Pet:
      description: Pet object that needs to be added to the store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Pet'
        application/xml:
          schema:
            $ref: '#/components/schemas/Pet'
    UserArray:
      description: List of user object
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/User'
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header